yatube/requirements.txt
atomicwrites==1.4.1
attrs==19.3.0
beautifulsoup4==4.11.1
black==22.10.0
certifi==2019.9.11
chardet==3.0.4
click==8.1.3
colorama==0.4.5
Django==2.2.19
django-debug-toolbar==3.2.4
Faker==12.0.1
flake8==5.0.4
idna==2.8
importlib-metadata==1.5.0
mccabe==0.7.0
mixer==7.1.2
more-itertools==8.2.0
mypy-extensions==0.4.3
packaging==20.1
pathspec==0.10.1
Pillow==6.2.0
platformdirs==2.5.2
pluggy==0.13.1
py==1.8.1
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==2.4.6
pytest==5.3.5
pytest-django==3.8.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
pytz==2019.3
requests==2.22.0
six==1.14.0
sorl-thumbnail==12.6.3
soupsieve==2.3.2.post1
sqlparse==0.3.0
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.4.0
urllib3==1.25.6
wcwidth==0.1.8
zipp==2.2.0
